% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlr_functions.R
\name{fill_matrix}
\alias{fill_matrix}
\title{Create a correlation matrix, checking for positive definite}
\usage{
fill_matrix(size = 6, cors = c("high", "mid", "low"), check.pd = TRUE)
}
\arguments{
\item{size}{Number of variables in the variance/covariance matrix}

\item{cors}{The size of the correlations between variables. Current options
include "high", "mid", and "low". Future iterations may include "toeplitz" and maybe
other options}

\item{check.pd}{Should the matrix be checked for a positive definite matrix? Defaults to TRUE. 
Probably should be FALSE if the number of variables is high.}
}
\value{
A randomly-generated correlation matrix
}
\description{
Create a correlation matrix, checking for positive definite
}
\examples{
fill_matrix(size=5, cors="high")
fill_matrix(size=25, cors="high", check.pd=FALSE)
}
