\name{qlr-package}
\alias{qlr-package}
\alias{qlr}
\docType{package}
\title{
\packageTitle{qlr}
}
\description{
\packageDescription{qlr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qlr}
\packageIndices{qlr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{qlr}

Maintainer: \packageMaintainer{qlr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
