% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlr_functions.R
\name{random_cor_cov}
\alias{random_cor_cov}
\title{Generate a random correlation (or covariance) matrix}
\usage{
random_cor_cov(size = 6, cors = c("high", "mid", "low"), sds = NULL)
}
\arguments{
\item{size}{The number of variables. Defaults to 6.}

\item{cors}{The size of the correlations between variables. Current options
include "high", "mid", and "low". Future iterations may include "toeplitz" and maybe
other options}

\item{sds}{A vector containing the standard deviation of each of the variables. 
Defaults to NULL, in which case, it randomly generates standard deviations that range
from .2 to 10}
}
\value{
A list containing
\itemize{
  \item cor - the randomly generated correlation matrix
  \item sd - a vector of standard deviations
  \item cov.mat - the variance/covariance matrix
}
}
\description{
Generate a random correlation (or covariance) matrix
}
\examples{
random_cor_cov(size=4, cors="low")
random_cor_cov(size=3, cors="mid", sds = c(5,5,5))
}
