% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlr_functions.R
\name{return_cor_limits}
\alias{return_cor_limits}
\title{Return the limits of a correlation}
\usage{
return_cor_limits(cors = c("high", "mid", "low"))
}
\arguments{
\item{cors}{The size of the correlations between variables. Current options
include "high", "mid", and "low". Future iterations may include "toeplitz" and maybe
other options}
}
\value{
A vector containing the ranges of the correlations
}
\description{
Return the limits of a correlation
}
